+++ R/geom-boxplot.R
@@ -118,6 +118,7 @@ geom_boxplot <- function(mapping = NULL, data = NULL,
                          outlier.alpha = NULL,
                          notch = FALSE,
                          notchwidth = 0.5,
+                         staplewidth = 0,
                          varwidth = FALSE,
                          na.rm = FALSE,
                          orientation = NA,
@@ -151,6 +152,7 @@ geom_boxplot <- function(mapping = NULL, data = NULL,
       outlier.alpha = outlier.alpha,
       notch = notch,
       notchwidth = notchwidth,
+      staplewidth = staplewidth,
       varwidth = varwidth,
       na.rm = na.rm,
       orientation = orientation,
@@ -211,7 +213,7 @@ GeomBoxplot <- ggproto("GeomBoxplot", Geom,
                         outlier.fill = NULL, outlier.shape = 19,
                         outlier.size = 1.5, outlier.stroke = 0.5,
                         outlier.alpha = NULL, notch = FALSE, notchwidth = 0.5,
-                        varwidth = FALSE, flipped_aes = FALSE) {
+                        staplewidth = 0, varwidth = FALSE, flipped_aes = FALSE) {
     data <- check_linewidth(data, snake_class(self))
     data <- flip_data(data, flipped_aes)
     # this may occur when using geom_boxplot(stat = "identity")
@@ -275,8 +277,28 @@ GeomBoxplot <- ggproto("GeomBoxplot", Geom,
       outliers_grob <- NULL
     }
 
+    if (staplewidth != 0) {
+      staples <- data_frame0(
+        x    = rep((data$xmin - data$x) * staplewidth + data$x, 2),
+        xend = rep((data$xmax - data$x) * staplewidth + data$x, 2),
+        y    = c(data$ymax, data$ymin),
+        yend = c(data$ymax, data$ymin),
+        alpha = c(NA_real_, NA_real_),
+        !!!common,
+        .size = 2
+      )
+      staples <- flip_data(staples, flipped_aes)
+      staple_grob <- GeomSegment$draw_panel(
+        staples, panel_params, coord,
+        lineend = lineend
+      )
+    } else {
+      staple_grob <- NULL
+    }
+
     ggname("geom_boxplot", grobTree(
       outliers_grob,
+      staple_grob,
       GeomSegment$draw_panel(whiskers, panel_params, coord, lineend = lineend),
       GeomCrossbar$draw_panel(
         box,